Crate groom

Groom

Groom is primarily a Command-Line Interface (CLI) application for processing mustache templates, but it is also implemented as a library (crate) that can be integrated with other projects. Since the binary is essentially a wrapper around the publicly exposed functionality of the crate, documentation on using the binary is provided here in the API documentation.

Binary Usage

The following examples show using the binary (executable) from a command line. The binary should work on Windows, macOS, and Linux.

Examples

The following example will render the input template using the data in the data.yml file and writes the output to stdout.

$ groom -m data.yml template.mustache

The following example renders the input template by reading the data from stdin and writes the output to stdout.

$ cat data.yml | groom template.mustache

The following example renders the input template using the data in the data.yml file and writes the output to a file.

$ groom -m data.yml template.mustache output

Exit Codes

Code 	Reason

0 	Success, no error

1 	Failure, generic

2 	Failure, user input

3 	Failure, Input/Output (IO)

4 	Failure, rendering mustache template

5 	Failure, UTF8 encoding

6 	Failure, YAML decoding

Structs

Groom
A builder for running the application.

Enums

Error
The error type for groom-related operations and associated traits.

Type Definitions

Result
A specialized Result type for groom operations.