Struct groom::Groom

pub struct Groom { /* fields omitted */ }

A builder for running the application.

Methods

impl Groom

pub fn new() -> Groom

Creates a new application instance.

pub fn data<P: AsRef<Path>>(self, d: Option<P>) -> Self

Sets the data.

If the data is None, then stdin is used for the data.

pub fn output<P: AsRef<Path>>(self, o: Option<P>) -> Self

Sets the output.

If the output is None, then stdout is used for the output stream.

pub fn run<P: AsRef<Path>>(self, input: P) -> Result<()>

Runs the application.

This will consume the Groom and process the input template using the provided mapping and write to the output.